backend/pyproject.toml
[tool.poetry]
name = "arain-association-backend"
version = "0.1.0"
description = "Backend API for Arain Association Youth Wing Pakistan NGO Web App"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
motor = "^3.3.2"
pydantic = {extras = ["email"], version = "^2.5.0"}
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
httpx = "^0.25.2"
reportlab = "^4.0.7"
pandas = "^2.1.4"
openpyxl = "^3.1.2"
pillow = "^10.1.0"
aiofiles = "^23.2.1"
beanie = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
